arch: provide thread-local storage helper functions
Because the layout of the TLS area is architecture-specific, we provide
helper functions to get the required size and alignment of the TLS area,
as well as a help function that does the actual copying of the TLS
"master copy" according the layout required by the ABI.

Only the x86-64 implementation is complete and tested. Arm32 is pure
boilerplate. Arm64 has some rough guesses about how the layout should
look like, but is untested.

Signed-off-by: Florian Schmidt <florian.schmidt@neclab.eu>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>