lib/ukdebug: trace.h - introduce UK_FOREACH and UK_NARGS macros
The UK_FOREACH is the central macro in the all macro hell needed to
implement static tracepoints definitions. The UK_FOREACH takes up to 7
parameters, and applies specified function to each of them, separating
output with commas.

For example it converts:
    "char, int, long"  =>  "sizeof(char), sizeof(int), sizeof(long)"

UK_NARGS calculates how many parameters it has received.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>