lib/uksched: Make waiting queue act like actual queues
Waiting queues were used as stacks (LIFO) instead of queues (FIFO). The
last added thread was the first one to be woken up which violates the
fairness quality of waiting queues. This patch fixes that by adding the
last added thread at the end of the queue.

The pthread-embedded library provides a test which verifies this propery
- the rwlock6 test.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>