lib/vfscore: do not advance position for stdio
With the proper handling of uio for stdio files, vfs also advances the
position in the struct file. Which does not make any sense for console
(and some other classes of devices).

This patch adds another flags variable inside struct
vfscore_file. Using f_flags would add more mess in the code, because
most of the flags are defined in fcntl.h, which is external from
vfscore header.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Signed-off-by: Vlad-Andrei Badoiu <vlad_andrei.badoiu@stud.acs.upb.ro>
Reviewed-by: Florian Schmidt <florian.schmidt@neclab.eu>