lib/syscall_shim: generate provided_syscalls.h
The provided_syscalls.h will be containing prototypes for every
syscall present in the system.

It will be the responsibility of the libraries to tell syscall_shim if
certain syscall is provided. If syscall is available, it must be
listed in the makefile's variable 'UK_PROVIDED_SYSCALLS' in the format
<syscall_name>-<number_of_arguments>. For example:

UK_PROVIDED_SYSCALLS-$(CONFIG_LIBVFSCORE) += writev-3

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>