lib/syscall_shim: introduce uk_syscallp
The macro layer is efficient, but the exact syscall must be known at
the compilation stage. If a syscall number is kept in a variable macro
layer will not be able to resolve it to the syscall function.

For this case another file is generated - syscall_entry.c. Where
syscalls are resolved at run-time into a destination functions.

Signed-off-by: Yuri Volchkov <yuri.volchkov@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>