lib/ukboot: Fix handling of main arguments
When the command line is parsed and split into arguments, each argument is saved
in an array of pointers allocated on the stack ('argv' array in
ukplat_entry_argp()). This means that the array can be overwritten anytime
before starting the main thread which needs the arguments. This patch fixes that
by making the array static.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>