Fix dentry referecing and file closing
Increase the refcount of the vfscore_file's dentry when opening it.
This is the OSv behaviour since each vfscore_file has an intrusive_ptr
ref to its dentry, which increases dentry's refcount on creation and
decreases dentry's refcount on deletion. We must increase the dentry's
refcount since we might end up with files pointing to deleted dentries,
if the actual file/directory the dentry refers gets deleted (so the
dentry's refcount hits 0 and it gets freed).

We must also decrease the refcount of the dentry on file closing.

The statement "Dentry stays forever in the dentry cache. Unless the
file/directory it refers to gets deleted/renamed" doesn't hold for
file systems that do not save their structures on permanent mediums
(like LWIP's sockets abstraction) so the dentry refcount is equal to
the number of vfscore_file's pointing at it (so the dentry gets
deleted as soon as the last file is closed).

Right now the vfscore_file is never freed, excepting the LWIP socket's
file which is freed inside sock_net_close along with its dentry and
vnode. This should not be the expected behaviour from a underlying
file system (ramfs provides a good example of expected behaviour).
As such, we must free the vfscore_file when closing it and release its
dentry so the dentry and vnode freeing will be handled by vfs as well.

This patch also makes fdrop return 0 if the file's refcount hits
0 or return 1 otherwise. This is the OSv behaviour and the return
of an error from fdrop is never checked.

Signed-off-by: Mihai Pogonaru <pogonarumihai@gmail.com>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>