build: Introduce UK_DEBUG_IMAGES, UK_DEBUG_IMAGES-y
UK_DEBUG_IMAGES, UK_DEBUG_IMAGES-y are introduced in order to enable
for each platform to specify which images contain debugging
information and for which ones gdb helper scripts should be generated.

Note, this patch temporarily disables the generation of those helper
scripts because none of the platforms populates the image names. The
follow-up patches are adopting the platform linker files in order to
re-introduce this functionality.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>