lib/posix-libdl: Introduce POSIX libdl library skeleton
A great number of libraries use explicit loading of dynamic libraries and for
this reason we provide the libdl Unikraft implementation. Given that libdl is
POSIX, we propose a new name convention for internal libraries which implement
POSIX functions: every internal library will have the 'posix-' prefix in order
to differentiate it from libraries that implement Unikraft-specific
functionality.

Signed-off-by: Andrei Gogonea <andrei.gogonea09@gmail.com>
Signed-off-by: Teodora Serbanescu <teo.serbanescu16@gmail.com>
Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>