plat/kvm: Fix current thread retrieval in interrupt context on x86_64
Commit 017fffd5 introduced support for setting the current thread pointer on top
of interrupt stacks in order to retrieve the current thread in interrupt context
as well. Unfortunately, the wrong stack was picked for KVM platform. This patch
fixes that and sets the thread on cpu_intr_stack instead.

cpu_intr_stack was resized to STACK_SIZE because this is a mandatory condition
when saving threads on top. However, given that it also needs a STACK_SIZE
alignment, a new section was created for it, .intrstack, in order to avoid
breaking the entire binary image layout. Without this new section, the entire
.text section would have a STACK_SIZE alignment (i.e. 64KB) and this would imply
that the multiboot header, which is included in .text section, would also be
moved at an address higher than STACK_SIZE, even though it must stay in the
first 8KB of the binary.

We apply the same solution for the traps stack in order to be able to print
the entire stack trace on faults.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>