plat/common: Add uk_ctortab_end in the linker script
Currently we use LONG(0) in the linker script and a NULL chec to find the end of
the ctortab section. This approach does not work on different arhitectures
because LONG(0) has only 32 bits. To make the script arch independent, we add a
new variable uk_ctortab_end at the end of the ctortab section.

Signed-off-by: Vlad-Andrei Badoiu <vlad_andrei.badoiu@stud.acs.upb.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>