lib/uktime: Introduce time functions library skeleton
Until now, the time functions were added one by one in both nolibc and newlib
glue code. This meant that one had to make the exact same changes in both
Unikraft and newlib repos. The major drawback is that one can forget to
duplicate the changes in the other library and this has already happened
several times. Now, given that both nolibc and newlib miss these time
functions, it makes more sense to provide an internal library for all of them -
i.e. keeping them in one place instead of more.

Some of the functions are implemented from scratch, but some may be imported
from other libraries. For those imported from musl we will keep the same
hierarchy under the musl-imported/ subdirectory, just like we do it for newlib
glue code. That's why we also add the musl COPYRIGHT.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Vlad-Andrei Badoiu<vlad_andrei.badoiu@stud.acs.upb.ro>