build: Introduce DBGFLAGS, DBGFLAGS-y
The original problem is that assembly source files are compiled
without debugging information. GCC's `-g[0-3]` flag is not handed over
to the assembler build rules. This causes the debugging information
of the final image to be incomplete.
Unfortunately the debug option flag cannot be simply added to
`ASFLAGS` because the linker script build rule uses them, too. Those
rules use GCC's preprocessor only but the debug flag causes GCC to
dump the internal preprocessor state into the generated output text
file.
Because all of this, this patch introduces the build system variables
`DBGFLAGS` and `DBGFLAGS-y`. They are intended to carry debug-related
compile flags (like `-g[0-3]`). They are handed over to only to the
AS, C, C++, and GO compile rules.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>