lib/vfscore, lib/*fs: Ignore incompatible function pointer typecast
Vnode operations are implemented with function calls. In order to stub
some of the operations a no-op function is filled to those struct
entries. Such a stub function just returns an error code that tells
that the functionality is not implemented. However, the function
signature may not match with the stubbed vnode operation, so it is
typecasted. GCC 8 introduces a warning when function pointer typecasts
are incompatible. This patch adds a CFLAG to ignore this warning for
vfscore and all current filesystem implementations.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>