lib/ukswrand: Add getrandom() function
This function is used for filling buffers with random data. We
introduce it now because it is used by the Python 3
interpreter. Although we define the GRND_NONBLOCK and GRND_RANDOM
flags, they are not used internally because we use the same "source"
as the uk_swrand_randr() function. Given that reading from
'/dev/random' implied the same core logic, we also refactor a bit and
use a common internal function for filling buffers:
uk_swrand_fill_buffer().

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>