Makefile: Include internal libs Makefile.uk before external libs Makefile.uk
The reason is that the internal libs headers should have priority because they
provide the definitions on which the external libs should depend.

Let's take the uktime internal lib for example. It provides the time related
definitions which should be used by all external libs and apps. When including
<time.h> or <sys/time.h>, this is the lib where these headers should come from.
If customizations or extensions are allowed in the external libraries then the
internal headers should enable this by using #include_next directives (see next
patch).

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>