lib/uktime: Introduce time_types.h for type defintions
We basically take the type definitions we used for nolibc and put them in
time_types.h.

In this patch we also add the changes for allowing custom definitions in libc
implementations. We allow for all libc implementations but for nolibc; the
alternative would have been to add empty headers in nolibc.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>