plat/xen/drivers/blk: Use pool of grant refs for each queue
Each read/write request needs a number of grant references. In order to avoid
allocating new grant refs at the beginning of every operation and freeing them
at the end of it, we reuse the same grant references stored in a pool. In case
the pool does not contain enough grant refs for the request, new ones are
allocated, and they are freed when the response is processed.

Signed-off-by: Roxana Nicolescu <nicolescu.roxana1996@gmail.com>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>