include/uk: Introduce <ctors_prio.h> header for constructors priority values
We introduce the <ctors_prio.h> header as a way for centralizing all the values
for constructor priorities. Whenever we will want to change the values we will
do it here.

It's not obvious from the beginning which values we should assign to the
constructors in the system, so we expect this file to be prone to many changes
in the future.

This is only a temporary solution until we will find a way to set the
priorities values for each application.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Vlad-Andrei Badoiu <vlad_andrei.badoiu@stud.acs.upb.ro>