lib/posix-process: Implement getpriority() and setpriority()
We allow getting and setting priority only for the calling process (or its
process group or its real user ID), meaning that 0 is the only accepted value
for the 'who' parameter. This patch also introduces the default value for the
Unikraft "process" priority. This is also the only value that can be set using
setpriority() (i.e. basically it doesn't change anything).

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>