plat/*: Add binary system call feature and API
Introduces the `HAVE_SYSCALL` feature flag that enables handling of
binary system calls. It is intended that the platform libraries are
enabling and implementing system call entrance for binary system call
requests when this flag is set.
This patch also introduces an extension to the platform API
(`include/uk/plat/syscall.h`) that defines a system call handler
prototype. It is intended that this handler should be implemented by a
non-platform library. A desing principle is that different system call
ABIs could be with this function signature. For this purpose, the trap
handler in the platform should save the CPU registers state and
hand-over a reference to the handler. The handler modifies this saved
state accordingly to the ABI standard. On returning, the trap should
then load the updated register values to the CPU and continue execution.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Gaulthier Gain <gaulthier.gain@uliege.be>