lib/vfscore: dup2(): Close newfd when already opened
If newfd is open, then POSIX says we have to close it - that's what this
patch aims to achieve. Given that dup2/dup3 explicitly wants to reserver a
file descriptor, we also introduce the vfscore_reserve_fd() function. We
also add a tiny fix in vfscore_alloc_fd() because it should also be able to
allocate file descriptor 0 - that's what stdio initialization needs among
other things.

Signed-off-by: Costin Lupu <costin.lupu@cs.pub.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>