include/uk: Remove <ctors_prio.h> header
This reverts commit d56a47c2d123c5e46c35d759be1b7dd9bc90b093 and
removes <uk/ctors_prio.h>.
The intention of the author was that constructors of Unikraft system
libraries use the definitions in this header to coordinate the order
of initialization. However, constructors defined with the constructor
attribute end up in the init_array and preinit_array. We learned from
recent porting works, that these tables need to be treated as being
application-specific and should not be used for Unikraft system
components. System libraries (e.g., vfscore) need to be initialized
earlier because the application may depend on them. The problem is
that our priority levels are possibly in conflict with the application
ones. For this purpose, we recently introduced uk_ctortab and
uk_inittab. Those tables are handled before the application constructors
and are intended for initializing Unikraft system components.

So far, only one external library used the scheme that is reverted
with this patch: pthread-embedded. Another patch is sent out to adopt
this library to the uk_inittab, uk_ctortab scheme.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>