Temporary workaround to have atomic operations working on ARM64 baremetal
At the moment atomic operations are working on ARM64 for KVM, however,
they are not working on baremetal on ARM64 (i.e., on the Raspberry Pi 3B+).
This can be most likely be solved by enabling the
caches and setting up a compatible memory configuration.
In the mean time, this compile guard allows us to have functionally-equivalent
non-atomic implementations of the gcc builtins working on baremetal on ARM64.
Once the issue is fixed, then this workaround can be removed.

Signed-off-by: Santiago Pagani <santiago.pagani@neclab.eu>