lib/posix-sysinfo: glibc version workaround for `uname` (prepend 5-)
During glibc initialization, the library is doing a `uname` system
call in order to detect the kernel version. It looks into the release
field and parses the string. We prepend (for us meaningless) "5-" to
our version string in order to tell glibc that it runs on a new enough
kernel.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Gaulthier Gain <gaulthier.gain@uliege.be>