plat/common/arm/generic_timer.c: Fix time conversion functions
The functions for time conversion uses multiplier and shift factors. However,
when the value of ticks (or nanoseconds) becomes too high, the conversion loses
the precision. In fact, this mechanism is only useful for low values, so this
commit handles the conversion using basic operations (multiplication or
division) for high values.

Signed-off-by: Renê de Souza Pinto <Rene.deSouzaPinto@opensynergy.com>
Reviewed-by: Răzvan Vîrtan <virtanrazvan@gmail.com>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #175