plat/common/arm/generic_timer.c: Keep consistence of delay calculation
generic_timer_cpu_block_until() receives the amount of waiting time in
nanoseconds. However, it needs to convert the total waiting time into ticks in
order to program the timer interrupt. In this conversion,
ukplat_monotonic_clock() is called to record the current nanoseconds and then
generic_timer_get_ticks() is called after in order to calculate the total
difference in ticks. This commit introduces a new function to retrieve the time
in nanoseconds and in ticks at once in order to keep the consistence of the
calculation, so the current time in nanoseconds and in ticks will correspond
to the same period, i.e., when now_ns is assigned.

Signed-off-by: Renê de Souza Pinto <Rene.deSouzaPinto@opensynergy.com>
Reviewed-by: Răzvan Vîrtan <virtanrazvan@gmail.com>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #175