lib/ukalloc: change page allocation api to take a number of pages
Change palloc and pfree to take a number of pages instead of an order as
argument.

Number of pages is then converted to an order within the binary buddy
allocator.  This can be done in constant time on many CPU architectures,
including x86 (rounding to log of the next power of two).

This allows for a cleaner interface and simplifies the implementation of
new, non-power-of-two-based memory allocators.

This also makes uk_malloc_ifpages more efficient, because we replace
uk_alloc_size_to_order which was basically an inefficient way of rounding
to log of the next power of two (assuming it wasn't optimized by a smart
compiler).

Port all calls to palloc within the main tree to the new API.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>