include: Clean-ups to <uk/ctors.h>
This commit brings the Unikraft constructor definitions closer to the
one of the Unikraft Init table.

UK_CTOR_FUNC() is basically replaced by UK_CTOR_PRIO(). Naming and
argument order is derived from Init table definition <uk/init.h>.

The table iterator is now similar as the one found in the Init
table. Instead of providing an integer as index, the foreach-macro is
directly iterating with a function pointer.

Because of compatibility reasons, the previous macro UK_CTOR_FUNC() is
still provided. The idea is that it gets removed as soon all depending
code was updated (internal and external).

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Vlad-Andrei Badoiu <vlad_andrei.badoiu@stud.acs.upb.ro>