arch/arm: Copy atomic.h from Arm64 to Arm32
Copies atomic.h from Arm64 to Arm32. The functions implemented in this
header are implemented based on GCC builtins. The architecture does
not matter for those functions, so the header works on both
architectures.
As long as we did not start restructuring arch/ and plat/, copying the
header is the most feasible option. Longer-term, the header could
become generic.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>