lib/vfscore: vfscore_put_fd: do not reject fd <= 2
Rejecting fd <= 2 is legitimate in some way, since freeing
std(in|out|err) might lead to crash if .f_count is 1. This is because
stdin, stdout and stderr fps are statically allocated.  However,
returning -EBUSY in this case breaks dup2 with stdin, out, and err.
Ignoring this should be fine as long as those are not fdrop-ed twice.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>