lib/vfscore: fix invalid error code
vfscore_vget returns 1 if the vnode was found in cache, 0 otherwise.

If we fall back to allocating a new vnode (vn_lookup returned NULL)
then the vnode was not found in cache and it does not make sense to
return anything else than 0.

In particular, this line is reached if VFS_VGET fails, meaning that
error will systematically be > 0. Since most calls to vfscore_vget
check for > 0 instead of == 1, they will assume that the vnode was
found in cache and dereference vpp without previously checking it,
causing a NULL pointer dereference.

This is not an issue for the moment since all vfs_vget implementations
are linked to vfscore_nullop, but might become an issue in the future
when those will be implemented.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@neclab.eu>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>