arch/*, build: Introduce reserved `isr` variant
In order to compile functions that can be called within interrupt
context, we introduce the reserved `isr` variant. As soon as this
variant is given to a source file, different ARCHFLAGS(-y) are used
for the compilation: ISR_ARCHFLAGS(-y). These flags are defined in the
architecture and forbid the compiler to use enhanced CPU functions
that are not saved within interrupt context (e.g., floating point
units, vector units).

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>