lib/syscall_shim: `UK_(LL)SYSCALL_R_DEFINE()`: Use given return type
Instead of declaring the return type for `UK_(LL)SYSCALL_R_DEFINE()`
functions as `long`, the user-given return type is used. In general,
this is done for convenience. Whenever a system call implementation
uses pointers as return type, the helpers defined in `<uk/errptr.h>`
can be used to encapsulate error codes.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Gaulthier Gain <gaulthier.gain@uliege.be>