lib/syscall_shim: Option to generate libc-style stubs
Introduces the option to automatically generate libc-style system call
stubs for unavailable system calls. With this function, it is possible
to provide all libc-style system call symbols although just a subset
of the full API is implemented. The symbols are defined as `weak` and
can be replaced by non-weak symbol definitions.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Gaulthier Gain <gaulthier.gain@uliege.be>