build: Support multiple symbol export files
Adds support such that `$(LIBNAME_EXPORTS)` can contain a list of symbol
export files instead of only a single one. In order to simplify
writing of `Makefile.uk` files, a `$(LIBNAME_EXPORTS-y)` variant is
provided, too. With this variant, export symbols can be dependent on
library configuration, for example:
 LIBNAME_EXPORTS-$(CONFIG_MYLIB_EXTENDEDABI) += extendedsyms.uk
Having this feature, library symbols can be masked even without the
existence of an `exportsyms.uk` file: Because of this, this
patch removes the "missing exportsyms.uk" warning. In case an
`exportsyms.uk` file exists for a library, it is still automatically
included and used: the previous behavior is preserved.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>