build: Provide option to localize symbols (LIBNAME_LOCALS(-y))
Similar to exporting symbols, this patch introduces the option to
localize symbols. The file format of a localizing symbol files is the
same as for exporting symbols (e.g., `exportsyms.uk`: Plain-text, one
symbol name per line. Such files can be registered to
`$(LIBNAME_LOCALS)` and `$(LIBNAME_LOCALS-y)`. As soon as a
`localsyms.uk` file exists in a library import base, it is
automatically added to the list `$(LIBNAME_LOCALS)` (similar behavior
as the handling of `exportsyms.uk files).

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Felipe Huici <felipe.huici@neclab.eu>