lib/vfscore: Fix double refcounting bug
This patch removes the `dref()` line in `sys_open` which would result
in an unbounded growth of the refcount. open() would take two
references: one via dentry_lookup() in namei(), and the other via
dref(); close() would release only one reference via drele().

As to the comment also deleted in this patch, it is true that OSv uses
intrusive_ptr, however it is created with the second argument
add_ref = false, hence the "std::move()" comment below in the original
OSv source code.

Now close() correctly releases the dentry if there are no open files
left.

Signed-off-by: Cristian Banu <cristb@gmail.com>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>