lib/uk9p: Recycle uk_9preq allocations
Each 9p device keeps a free-list of 9p requests. This allows
allocations to take much less time, as usually the workflow of a 9p
device is synchronous: create request, wait for reply, process it,
repeat. This approach avoids the allocation performance hit on
subsequent request creations.

Results below indicate that request_allocated takes < 100ns,
as opposed to 1us (previous commit in this patch series).

In the sample below, the read() now spends most of the time (~99.8%)
either notifying the host or waiting for a reply. At this scale,
even uk_traces are probably noisy and the "real" time might be lower
by a relatively large amount.

Excerpt from the uk traces on a read():

500281499  uk_9p_trace_request_create
500281570  uk_9p_trace_request_allocated
500281603  uk_9p_trace_ready              tag 0
500283423  uk_9p_trace_sent               tag 0
500347243  uk_9p_trace_received           tag 0

Signed-off-by: Cristian Banu <cristb@gmail.com>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>