lib/9pfs: Simplify the logic for removes
The logic for unlinking is simplified, either:
- there are no open files, remove now;
- there are open files, defer until all open files are closed

This is now possible since successive open and closes update
the dentry reference count correctly, so VOP_INACTIVE is called
in time to free the fid before the qid starts being reused
on the server-side (and we get an inode collision).

This also means that on unmounting, releasing all non-root
dentries and calling vfscore_release_mp_dentries() for m_covered
and m_root is enough to ensure all fids are clunked before
disconnecting.

Signed-off-by: Cristian Banu <cristb@gmail.com>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>