Remove default value
Setting a specific value to a choice config (from one of the libraries) cannot
overwrite a default value (that is set in Unikraft). It seems this is the case
for our C based Kconfig (from my knowledge we use this) - a default value set
earlier (in Unikraft) has a higher priority than a one set later (in a library).

For solving this issue, one solution is to use "named choices" - did that for
the tensorflowlite porting.

The information is obtained from an issue discussed here[1].

[1] https://github.com/zephyrproject-rtos/zephyr/issues/6948

Signed-off-by: George Muraru <murarugeorgec@gmail.com>
Reviewed-by: Hugo Lefeuvre <hugo.lefeuvre@neclab.eu>