plat/virtio_9p: Remove excessive check for max virtqueue size
For qemu virtio mmio read with VIRTIO_MMIO_QUEUE_NUM_MAX, it will
unconditionally return VIRTQUEUE_MAX_SIZE(1024) unlike virtio pci.
But it can set the number of queue size when seting up the virtqueue.
See vm_setup_vq which will be fixed in coming patch.

So the excessive check in virtio 9p is not necessary.

Signed-off-by: Jia He <justin.he@arm.com>