plat/*: pci_bus: implement recursive PCI bus enumeration
The PCI bus was previously enumerated in a brute-force manner, scanning
all device on all buses (256 buses and 32 devices per bus, i.e., 8192
devices in total), resulting in tens of thousands of I/O read and write
operations at startup (~55ms overhead).

This patch implements recursive PCI bus enumeration: we first enumerate
the first bus (or the first bus*es* in the case of multiple PCI host
controllers), and recursively probe further buses as we discover further
PCI-to-PCI bridges. This reduces the overhead to ~1-2ms.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@neclab.eu>
Reviewed-by: Daniel Dinca <dincadaniel97@gmail.com>