arch: Define the cache line size for each architecture
This definition statically describes the d-cache line size in
bytes for {x86_64,arm,arm64}.

The d-cache line is a L1 data cache nearest to the processor and
the line size refers to the data memory block.

The d-line cache size can also be discovered by querying the
d-cache build egister at the expense of using more memory and
instruction cycles during boot.  Defining this size statically
increases performance since it is already known for the
respective architecture.

Signed-off-by: Alexander Jung <alexander.jung@neclab.eu>
Reviewed-by: Costin Lupu <costin.lupu@cs.pub.ro>