lib/ukallocregion: add region-based allocator
Add ukallocregion, a minimalist region-based allocator.

Note that deallocation is not supported. This makes sense because regions
only allow for deallocation at region-granularity. In our case, this would
imply the freeing of the entire heap, which is generally not possible.

Obviously, the lack of deallocation support makes ukallocregion a fairly
bad general-purpose allocator. This allocator is interesting in that it
offers maximum speed allocation and deallocation (bounded O(1), no
bookkeeping). It can be used as a baseline for measurements (e.g., boot
time) or as a first-level allocator in a nested context.

Refer to Gay & Aiken, `Memory management with explicit regions' (PLDI'98)
for an introduction to region-based memory management.

Signed-off-by: Hugo Lefeuvre <hugo.lefeuvre@neclab.eu>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>