lib/ukalloc: add ifmalloc compatibility interface
Add ifmalloc, the malloc compatibility interface. This interface
implements a POSIX compliant allocation interface on top of a simple
malloc() and free() interface. This interface is similar to ifpages,
which does the same for palloc() and pfree().

ifmalloc will be used for the port of TLSF and tinyalloc which do not
support (posix_)memalign().

In order to keep implementations consistent, this patch also addresses
a bug in uk_posix_memalign_ifpages() which did not handle size = 0
correctly. The function previously returned an error code and set
memptr to NULL, but POSIX allows only one of the two.  Both
implementations now return EINVAL and leave memptr untouched.

Signed-off-by: Hugo Lefeuvre <hle@owl.eu.com>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>