lib/uknetdev: Provide IEEE 802.1q, 802.1ad, Jumbo frame definitions
This commit includes support for IEEE 802.1q and IEEE 802.1ad tagged
frames to the recently introduced Ethernet header and frame definitions.
Due to these standards, an Ethernet frame (without FCS/CRC) can be 1522 B
(non-jumbo) and 9022 B (jumbo).
Together with this commit, the definitions are name spaced with `UK_`
prefix. This is done to avoid possible macro name clashes when porting
network stacks or packet processing frameworks to libuknetdev.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>