lib/syscall_shim: Force evaluation order of syscall arguments
The file gen_uk_syscall.awk is used to generate the file uk_syscall.c.
The generated code calls the syscall handlers like this:

uk_syscall_e_fstat(va_arg(arg, long), va_arg(arg, long))

This is a problem since the order in which these va_arg expressions are being
evaluated is unspecified. This leads to arguments being swapped, depending
on how gcc decides to evaluate this.

When printing to stdout from musl this leads to an endless loop in
__stdio_write.

This patch fixes this by introducing temporary variables to store the varargs
and then passing these variables to the syscall handlers, taking into account
the number of required arguments.

Signed-off-by: Felix Moebius <felix.moebius@posteo.de>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>