solve ukswrand does not generate random numbers
If you configure ukswrand with the platform timestamp
as initial seed, and you'll try to use getrandom function
or try to read from /dev/urandom you won't get random
numbers. This is because the initialization of the random
number generator, which is located in the swrand.c file,
is done earlier than the initialization of TSC clock.

Signed-off-by: ARGINT DRAGOS IULIAN <dragosargint21@gmail.com>
Reviewed-by: Laurentiu Barbulescu <lrbarbulescu@gmail.com>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #71