plat/xen: netfront: Simplify tansmit buffer GC function
Simplifies the garbage collection function for transmit
buffers (`network_tx_buf_gc()`): Basically the loop
that checks for more slots to clean-up
(`RING_FINAL_CHECK_FOR_RESPONSES`) is moved to the outside.
THe intention is that the GC function could be called in
contexts where some buffers should be garbage collected
but a full clean-up is not required (e.g., before
transmitting on a full queue).

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>