lib/vfscore: Change preadv and pwritev fd error handling.
This commit discards the `has_error` method in favour of in-line checks against
the file descriptor.  In BSD's internal implementation, the error is simply
passed upwards from `dofilewrite` and `dofileread`.  Additionally, the
`has_error` method would return a false positive in the circumstance where an
error was non-zero and did not constitute a EWOULDBLOCK or EINTR and the number
of bytes read was greater than zero.  This scenario was discovered when a
buffer greater than the number of bytes available was read from a vnop read op
which returns the number of bytes read (as per POSIX requirements).

Additional checks inline with FreeBSD's[0][1] implementation are for whether
the fd is seekable; the offset of the file is non-zero; and, the file type is
not a character device.

[0]: https://github.com/freebsd/freebsd/blob/6af4a8cb884975f65a51b3b6d766e084faf55d4f/sys/kern/sys_generic.c#L320
[1]: https://github.com/freebsd/freebsd/blob/6af4a8cb884975f65a51b3b6d766e084faf55d4f/sys/kern/sys_generic.c#L522

Signed-off-by: Alexander Jung <alexander.jung@neclab.eu>