plat/drivers/rtc: Enable pl031 as RTC for ARM64
Use pl031 as RTC for ARM64. Currently this provides a pl031-specific
implementation with an initial skeleton for generic RTC support.

The pl031-specific interface allows the reading / writing of time and
the setting up of alarms to be delivered as interrupts.

Initialization relies on information from the device tree blob (DTB).

Signed-off-by: Wei Chen <wei.chen@arm.com>
Signed-off-by: Jianyong Wu <jianyong.wu@arm.com>
Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Reviewed-by: Sharan Santhanam <sharan.santhanam@neclab.eu>
Reviewed-by: Michalis Pappas <mpappas@fastmail.fm>
Reviewed-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Approved-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #316