lib/isrlib: "Interrupt-service-routine"-safe standard functions
Introduces libisr, a library for providing "interrupt-service-routine"-
safe variants of standard functions so that they can be used in
interrupt handlers and early boot code. This commit provides an initial
set of standard functions normally found with `<string.h>`. The
implementation is derived from nolibc and compiled with the `|isr`
variant which disables the use of extended CPU features, like
vector units (e.g., SSE, AVX) and floating point units.

Signed-off-by: cristian-vijelie <cristianvijelie@gmail.com>
Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>