Correct license headers introduced by UPB
This commit removes the additional clause
"THIS HEADER MAY NOT BE EXTRACTED OR MODIFIED IN ANY WAY"
from BSD license headers that got introduced by mistake
by Cristian Banu with the following commits:
 ca9616f9a - lib/uk9p: Add 9P fid abstraction
 b2ce99adc - lib/uk9p: Add 9P transport registration
 5c24de843 - lib/uk9p: Add 9P device implementation
 440259b39 - plat/xen/drivers/9p: Initialize xenbus device
 5282598ae - plat/drivers: Virtio 9P driver skeleton
 7ee9a512d - plat/xen/drivers: Add 9pfront driver skeleton
 f2c29a657 - lib/uk9p: Add 9P requests abstraction
 f74733d49 - lib/9pfs: Implement mount and unmount
 cd65e50e4 - lib/uk9p: Add 9P protocol core definitions
 ef1021378 - lib/9pfs: Add 9pfs header
 e77b99c78 - lib/uk9p: Add 9P helper API

On the one hand, this additional clause is redundant because
the BSD license already states that it must remain and the
copyright notice must be kept.
On the other hand, the clause freezes the file header
and prohibits future contributors from extending the
copyright notice for their contributions. This additional
clause is not part of the official BSD 3-clause.

The original author consents to the change with a `Signed-off-by`
or `Reviewed-by` tag to this commit.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro> on behalf of Cristian Banu